refatoring: dataset telefonia_movel
