+++ pipelines/utils/crawler_anatel/telefonia_movel/utils.py
@@ -5,23 +5,19 @@ General purpose functions for the br_anatel_telefonia_movel project of the pipel
 # pylint: disable=too-few-public-methods,invalid-name
 
 import os
-from io import BytesIO
-from pathlib import Path
-from urllib.request import urlopen
 from zipfile import ZipFile
 import time
-import pandas as pd
 from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from pipelines.utils.utils import log
-from pipelines.datasets.br_anatel_telefonia_movel.constants import (
+from pipelines.utils.crawler_anatel.telefonia_movel.constants import (
     constants as anatel_constants,
 )
 
 
-def download(path):
+def download_zip_file(path):
     if not os.path.exists(path):
         os.makedirs(path)
     options = webdriver.ChromeOptions()
@@ -71,8 +67,8 @@ def download(path):
     log("Terminou o time...")
     log(os.listdir(path))
 
-def descompactar_arquivo():
-    download(path = anatel_constants.INPUT_PATH.value)
+def unzip_file():
+    download_zip_file(path = anatel_constants.INPUT_PATH.value)
     # Obtenha o nome do arquivo ZIP baixado
     zip_file_path = os.path.join(anatel_constants.INPUT_PATH.value, 'acessos_telefonia_movel.zip')
     time.sleep(300)
@@ -83,65 +79,3 @@ def descompactar_arquivo():
 
     except Exception as e:
             print(f"Erro ao baixar ou extrair o arquivo ZIP: {str(e)}")
-
-def to_partitions_microdados(
-    data: pd.DataFrame, partition_columns: list[str], savepath: str
-):
-    """Save data in to hive patitions schema, given a dataframe and a list of partition columns.
-    Args:
-        data (pandas.core.frame.DataFrame): Dataframe to be partitioned.
-        partition_columns (list): List of columns to be used as partitions.
-        savepath (str, pathlib.PosixPath): folder path to save the partitions
-    Exemple:
-        data = {
-            "ano": [2020, 2021, 2020, 2021, 2020, 2021, 2021,2025],
-            "mes": [1, 2, 3, 4, 5, 6, 6,9],
-            "sigla_uf": ["SP", "SP", "RJ", "RJ", "PR", "PR", "PR","PR"],
-            "dado": ["a", "b", "c", "d", "e", "f", "g",'h'],
-        }
-        to_partitions(
-            data=pd.DataFrame(data),
-            partition_columns=['ano','mes','sigla_uf'],
-            savepath='partitions/'
-        )
-    """
-
-    if isinstance(data, (pd.core.frame.DataFrame)):
-        savepath = Path(savepath)
-
-        # create unique combinations between partition columns
-        unique_combinations = (
-            data[partition_columns]
-            .drop_duplicates(subset=partition_columns)
-            .to_dict(orient="records")
-        )
-
-        for filter_combination in unique_combinations:
-            patitions_values = [
-                f"{partition}={value}"
-                for partition, value in filter_combination.items()
-            ]
-
-            # get filtered data
-            df_filter = data.loc[
-                data[filter_combination.keys()]
-                .isin(filter_combination.values())
-                .all(axis=1),
-                :,
-            ]
-            df_filter = df_filter.drop(columns=partition_columns)
-
-            # create folder tree
-            filter_save_path = Path(savepath / "/".join(patitions_values))
-            filter_save_path.mkdir(parents=True, exist_ok=True)
-            file_filter_save_path = Path(filter_save_path) / "data.csv"
-
-            # append data to csv
-            df_filter.to_csv(
-                file_filter_save_path,
-                index=False,
-                mode="a",
-                header=not file_filter_save_path.exists(),
-            )
-    else:
-        raise BaseException("Data need to be a pandas DataFrame")
